healthdash/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-WTF>=1.2.0
APScheduler>=3.10.0
Requests>=2.31.0
gunicorn>=21.2.0
python-dotenv>=1.0.0